+++ bin/sozo/src/commands/options/account/mod.rs
@@ -151,6 +151,7 @@ impl AccountOptions {
 #[cfg(test)]
 mod tests {
     use clap::Parser;
+    use katana_runner::RunnerCtx;
     use starknet::accounts::ExecutionEncoder;
     use starknet::core::types::Call;
     use starknet_crypto::Felt;
@@ -211,8 +212,9 @@ mod tests {
         assert!(cmd.account.account_address(None).is_err());
     }
 
-    #[katana_runner::katana_test(2, true)]
-    async fn legacy_flag_works_as_expected() {
+    #[tokio::test]
+    #[katana_runner::test(accounts = 2, fee = false)]
+    async fn legacy_flag_works_as_expected(runner: &RunnerCtx) {
         let cmd = Command::parse_from([
             "sozo",
             "--legacy",
@@ -235,8 +237,9 @@ mod tests {
         assert!(*result.get(3).unwrap() == Felt::from_hex("0x0").unwrap());
     }
 
-    #[katana_runner::katana_test(2, true)]
-    async fn without_legacy_flag_works_as_expected() {
+    #[tokio::test]
+    #[katana_runner::test(accounts = 2, fee = false)]
+    async fn without_legacy_flag_works_as_expected(runner: &RunnerCtx) {
         let cmd = Command::parse_from(["sozo", "--account-address", "0x0", "--private-key", "0x1"]);
         let dummy_call = vec![Call {
             to: Felt::from_hex("0x0").unwrap(),
