+++ crates/katana/runner/Cargo.toml
@@ -12,7 +12,6 @@ katana-runner-macro = { path = "macro" }
 anyhow.workspace = true
 assert_fs.workspace = true
 chrono.workspace = true
-rstest.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 starknet.workspace = true
